backend/requirements.txt
amqp==2.5.2
asgiref==3.2.3
Babel==2.8.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
# djangorestframework-filters==0.10.2
django-model-utils==4.0.0
django-rest-swagger==2.2.0
django-timezone-field==4.0
djangorestframework==3.11.0
flower==0.9.3
idna==2.9
importlib-metadata==1.5.0
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.7
lxml==4.5.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.0.0
psycopg2-binary==2.8.4
python-crontab==2.4.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
redis==3.4.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
tornado==4.2.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
XlsxWriter==1.2.8
zipp==3.0.0
